fix race condition involving engine.destroy() and `ready` event
The 'ready' event sets up an interval timer that is cleared in
`destroy()`. However, in some situations, it is possible to for the
interval to be created after `destroy()` has run, resulting in an
interval timer that is never cleared, and thus an app that does not
exit when it is supposed to. This change fixes the issue and adds a
test for it.

Refs: https://github.com/mafintosh/torrent-stream/issues/198